git-svn-id: http://62.90.102.41/svn/NiceHouse/trunk@1912 1431c75d-6ac1-4f6d-bc36-6b0d284ea005

+++ Management/pdf.py
@@ -385,7 +385,7 @@ class MonthDemandWriter:
                 s.restore_date = self.demand.finish_date
                 c_final_new = s.c_final
                 diff = c_final_new - c_final_old
-                total += diff * s.price_final / 100
+                total += int(diff * s.price_final / 100)
                 row.extend([c_final_old, c_final_new, diff, commaise(diff * s.price_final / 100)])
                 row.reverse()
                 rows.append(row)
